requirements.txt
discord
discord.py[voice]
pyYAML
requests